python_ai/requirements.txt
google-generativeai
numpy
scikit-learn
pandas
